+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - **New**: Added messaging support for hosting application.
 - **Added**: Support for chart provider `defaultOpacity` property.
 - **Added**: Initial "time slider" support for WMS / WMTS Overlays with a time dimension.
+- **Updated**: Use `mmsi` path time as "Last updated" time for AIS targets (#303).