+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - **Added**: Support for chart provider `defaultOpacity` property.
 - **Added**: Initial "time slider" support for WMS / WMTS Overlays with a time dimension.
 - **Updated**: Use `mmsi` path time as "Last updated" time for AIS targets (#303).
+- **Updated**: Use VesselFinder (replaces MarineTraffic) to display vessel information (#305).